Refactor WhiteMovingBox
