+++ src/app/kontakt/ContactDojos.tsx
@@ -7,6 +7,8 @@ import Image from "next/image";
 import { Suspense, useState } from "react";
 import { FaSpinner } from "react-icons/fa";
 
+import { TwoElementMovingBox } from "@/components";
+
 import {
   ContactDojosTab,
   ContactAddress,
@@ -35,36 +37,6 @@ const dojoContentVariants = (toLeft: boolean) => {
   return variant;
 };
 
-interface WhiteMovingBoxProps {
-  selectedIndex: number;
-}
-
-export const WhiteMovingBox = ({ selectedIndex }: WhiteMovingBoxProps) => {
-  const whiteMovingBox: Variants = {
-    active: {
-      left: "0%",
-      transition: {
-        ease: "easeOut",
-        duration,
-      },
-    },
-    inactive: {
-      left: "50%",
-      transition: {
-        ease: "easeOut",
-        duration,
-      },
-    },
-  };
-  return (
-    <motion.div
-      variants={whiteMovingBox}
-      animate={selectedIndex === 0 ? "active" : "inactive"}
-      className="bg-white absolute inset-0 w-[50%] rounded-2xl"
-    />
-  );
-};
-
 export const ContactDojos = () => {
   const [selectedIndex, setSelectedIndex] = useState(0);
 
@@ -81,7 +53,7 @@ export const ContactDojos = () => {
         <ContactDojosTab selectedIndex={selectedIndex} index={1}>
           KSG Spordisaal Sikupillis
         </ContactDojosTab>
-        <WhiteMovingBox selectedIndex={selectedIndex} />
+        <TwoElementMovingBox selectedIndex={selectedIndex} />
       </Tab.List>
       <Tab.Panels>
         <AnimatePresence>
