+++ src/components/elements/index.ts
@@ -6,3 +6,4 @@ export * from "./Modal";
 export * from "./ResizablePanel";
 export * from "./user";
 export * from "./time-date-pickers";
+export * from "./moving-box";
