+++ src/components/elements/moving-box/index.ts
@@ -1,0 +1,1 @@
+export * from "./MovingBox";
